backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# RAG Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.16
openai>=1.10.0

# Vector DB
qdrant-client>=1.7.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
pillow>=10.0.0

# Text Processing
tiktoken>=0.5.0

# Utils
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0

# Monitoring
prometheus-client>=0.19.0

# Logging
python-json-logger>=2.0.0
